+++ server/interfaces/messages/TokensAssigned.ts
@@ -1,0 +1,6 @@
+export default interface TokenMessage {
+  entries: {
+    capture_id: string
+  }[]
+  wallet_name: string
+}
